Productionizing
Creating the Flask infrastructure. Was able to make curl requests to the wsgi.py file with the first row of test data and get a response (string, bc. int32 is not serilizable by JSON).

Couldn't create an average of train_test_split models --> had to pick a random seed --> chose one with acc, adjacency, and MAE close to the avg of 50 (production.py)